Split the things again
